% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_info_parser.R
\name{account_info_parser}
\alias{account_info_parser}
\title{Account data parser}
\usage{
account_info_parser(
  response,
  return_extra_attributes = TRUE,
  unfold_extra_attributes = FALSE
)
}
\arguments{
\item{response}{A response object as returned by the \code{httr}'s request verbs
(e.g., \code{GET} or \code{POST}).}

\item{unfold_extra_attributes}{}
}
\value{

}
\description{
Extracts information about an account from an API response object
}
\examples{

}
