% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{api_request}
\alias{api_request}
\title{Make an API call}
\usage{
api_request(url, max_attempts = 3L)
}
\arguments{
\item{url}{(character): URL to call.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A named list, the structure of which will vary depending on the
actual \code{GET} request made.
}
\description{
Performs \code{GET} requests, with an exponential backoff mechanism built in
}
\details{
This function only performs \code{GET} requests. The \code{url} is
expected to be built properly before passing onto this function.
The returned data are expected to be in JSON format. This function
will automatically parse the response object and return a named R list
with the respective elements.
}
\examples{
base_url <- "https://api.trongrid.io"
address <- "TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux"
url <- httr::modify_url(base_url, path = c("v1", "accounts", address))
url <- httr::parse_url(url)
url$query <- list(only_confirmed = tolower(TRUE))
url <- httr::build_url(url)

r <- api_request(url)
print(r)
}
