% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_address.R
\name{convert_address}
\alias{convert_address}
\title{Convert addresses}
\usage{
convert_address(address)
}
\arguments{
\item{address}{(character) Account address}
}
\value{
Account address (character). If \code{address} is in \code{hex} format, it
will be converted to \code{base58check} format, and the other way around.
}
\description{
Converts between \code{hex} and \code{base58check} formats of account addresses
}
\examples{
hex_address <- "41357a7401a0f0c2d4a44a1881a0c622f15d986291"
base58_address <- "TEqyWRKCzREYC2bK2fc3j7pp8XjAa6tJK1"
convert_address(hex_address)
convert_address(base58_address)

}
