% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_contract_type_id.R
\name{convert_contract_type_id}
\alias{convert_contract_type_id}
\title{Convert contract type ID}
\usage{
convert_contract_type_id(id)
}
\arguments{
\item{id}{(numeric or character): a vector of numeric or coercible-to-numeric
values to be mapped to \code{contractType} names. Cannot contain \code{NA} values.}
}
\value{
A charcter vector with \code{contractType} names.
}
\description{
Converts integer-encoded contract type IDs to human-readable names
}
\details{
Transaction attributes returned by the Tronscan API contain
\code{contractType}, an integer-encoded type of transaction. The dictionary
used by this function to map these integers to human-readable names
is taken from the official TRON Protocol
\href{https://github.com/tronprotocol/documentation/blob/master/English_Documentation/TRON_Virtual_Machine/TRC10_TRX_TRANSFER_INTRODUCTION_FOR_EXCHANGES.md}{documentation}.
}
\examples{
convert_contract_type(c(1, 2, 31))

}
