% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_unix_timestamp.R
\name{from_unix_timestamp}
\alias{from_unix_timestamp}
\title{Convert to POSIXct datetime values}
\usage{
from_unix_timestamp(ts, tz = "UTC")
}
\arguments{
\item{ts}{(character or numeric): Unix timestamps,
\emph{including milliseconds}.}

\item{tz}{(character): timezone of the returned datetime values.
Defaults to \code{UTC}.}
}
\value{
A \code{POSIXct} datetime value in the format \verb{\%Y-\%m-\%d \%H:\%M:\%S}.
}
\description{
Converts a Unix timestamp to a POSIXct datetime value
}
\examples{
ts1 <- 60 * 60 * 24 * 1000
ts2 <- as.character(60 * 60 * 24 * 1000)
from_unix_timestamp(ts1, tz = "UTC")
from_unix_timestamp(ts2, tz = "UTC")
}
