% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_balance.R
\name{get_account_balance}
\alias{get_account_balance}
\title{Get TRX and TRC-20 balances}
\usage{
get_account_balance(address, only_confirmed = FALSE, max_attempts = 3L)
}
\arguments{
\item{address}{A character value - address of the account of interest, in
\code{base58} (starts with \code{T}) or  \code{hex} (starts with \code{41}) format.}

\item{only_confirmed}{A boolean value. If \code{TRUE}, account balance will be
returned as of the latest confirmed block, otherwise as of the
latest unconfirmed one. Defaults to \code{FALSE}.}

\item{max_attempts}{A non-zero, positive integer specifying the maximum
number of attempts to call the API if the first attempt fails
(i.e. its call status is different from \code{200}). Additional attempts are
implemented with an exponential backoff. Defaults to 3.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{request_time}: date and time  (UTC timezone) when the API
request was made;
\item \code{address}: a character value indicating the account address
(in \code{hex} format);
\item \code{trx_balance}: a character value.
\item \code{trc20_balance}: a list that contains a tibble with two columns:
\code{trc20} (\code{base58}-formatted address of the token) and \code{balance}
(a character value, amount of the respective TRC-20 token).
}
}
\description{
Returns the current TRX and TRC-20 token balances of an account
}
\details{
In contrast to \code{get_trx_balance} and \code{get_trc20_balance}, this
function returns both the TRX and TRC-20 token balances.

All balances are presented with a precision of 6. This means
that a balance returned by this function needs to be divided by 6
(after converting to \code{as.numeric}) to obtain the actual value.
}
\examples{
r <- get_account_balance("TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux")
print(r)

}
