% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_info_by_address.R
\name{get_account_info_by_address}
\alias{get_account_info_by_address}
\title{Get account information by address}
\usage{
get_account_info_by_address(address, only_confirmed = FALSE, max_attempts = 3L)
}
\arguments{
\item{address}{A character value - address of the account of interest, in
\code{base58} or  \code{hex} format.}

\item{only_confirmed}{A boolean value. If \code{TRUE}, account details will be
returned as of the latest confirmed block, otherwise as of the
latest unconfirmed one. Defaults to \code{FALSE}.}

\item{max_attempts}{An integer specifying the maximum number of attempts
to call the API if the first attempt fails (i.e. its call status is
different from \code{200}). Additional attempts are implemented with an
exponential backoff. Defaults to 3.}
}
\value{

}
\description{
Returns account details based on its address
}
\examples{
address <- "TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux"

}
