% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_trx_balance.R
\name{get_account_trx_balance}
\alias{get_account_trx_balance}
\title{Get Tronix balance}
\usage{
get_account_trx_balance(address, only_confirmed = FALSE, max_attempts = 3L)
}
\arguments{
\item{address}{(character) - address of the account of interest, in
\code{base58} (starts with \code{T}) or  \code{hex} (starts with \code{41}) format.}

\item{only_confirmed}{(boolean) - if \code{TRUE}, account balance will be
returned as of the latest confirmed block, otherwise as of the
latest unconfirmed one. Defaults to \code{FALSE}.}

\item{max_attempts}{(integer, positive) - a non-zero integer, maximum
number of additional attempts to call the API if the first attempt fails
(i.e. its call status is different from \code{200}). Additional attempts are
implemented with an exponential backoff. Defaults to 3.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{request_time} (POSIXct, UTC timezone) - date and time when the API
request was made;
\item \code{address} (character) - address of the account, in \code{base58} format;
\item \code{trx_balance} (character) - the amount of Tronix
}
}
\description{
Returns the current Tronix (TRX) balance of an account
}
\details{
All balances are presented with a precision of 6. This means
that a balance returned by this function needs to be divided by
1 million (after converting with \code{as.numeric()}) to obtain the actual
amount of TRX.
}
\examples{
r <- get_account_trx_balance("TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux")
print(r)

}
