% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_info.R
\name{get_block_info}
\alias{get_block_info}
\title{Get block attributes}
\usage{
get_block_info(latest = TRUE, block_number = NULL, max_attempts = 3L)
}
\arguments{
\item{latest}{(boolean): indicator of whether the latest block's information
should be retrieved (\code{TRUE} by default).}

\item{block_number}{(character): number of the block to retrieve the
information for (\code{NULL} by default). If \code{latest = TRUE}, this argument is
ignored.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble with the following columns:
\itemize{
\item \code{block_number} (integer): number of the block;
\item \code{timestamp} (POSIXct): date and time when the block was created;
\item \code{hash} (character): hash ID of the block;
\item \code{parent_hash} (character): hash ID of the parent block;
\item \code{tx_trie_root} (character): hash ID of the
\href{https://en.wikipedia.org/wiki/Trie}{trie} root of the block creation
transaction;
\item \code{confirmed} (boolean): an indicator of whether this block has been
confirmed;
\item \code{size} (integer): size of the block (in bytes);
\item \code{witness_address} (character): address of the block's creator;
\item \code{tx_count} (integer): number of transactions associated with this block;
\item \code{tx} (list): a list with one element that contains a tibble with basic
info on transactions associated with this block: \code{tx_id} (character) -
transaction ID, \code{contract_type} (character) - type of the contract that
performed this transaction, \code{from_address} (character) - address of the account
that initiated the transaction, and \code{to_address} (character) - address of the
recieving account. All addresses are presented in the \code{base58check} format.
}
}
\description{
Retrieves information on a block and its transactions
}
\details{
Additional details on the block producing process can be found
in the \href{https://tronprotocol.github.io/documentation-en/introduction/dpos/#dpos}{official documentation}.
}
\examples{
# latest block
r1 <- get_block_info(latest = TRUE)
print(r1)

# specific block:
r2 <- get_block_info(latest = FALSE, block_number = "26810333")
print(r2)
}
