% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blocks_for_time_range.R
\name{get_blocks_for_time_range}
\alias{get_blocks_for_time_range}
\title{Get blocks for a time range}
\usage{
get_blocks_for_time_range(min_timestamp, max_timestamp, max_attempts = 3L)
}
\arguments{
\item{min_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
beginning of the period of interest (inclusive). Defaults to 0.}

\item{max_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
end of the period of interest (inclusive).}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{block_number} (character);
\item \code{timestamp} (POSIXct): date and time when the block was produced;
\item \code{hash} (character): hash ID of the block;
\item \code{parent_hash} (character): hash ID of the parent block;
\item \code{tx_trie_root} (character): \href{https://en.wikipedia.org/wiki/Trie}{trie} root
ID of the transaction that produced the block;
\item \code{confirmed} (boolean): indicator of whether the block is confirmed;
\item \code{revert} (boolean): indicator of the whether the block-producing
transaction was reverted;
\item \code{size} (integer): block size (in bytes);
\item \code{witness_address} (character): \code{base58check}-formatted address of the
account that produced the block;
\item \code{witness_name} (character): common name of the account that produced the
block;
\item \code{tx_count}(integer): number of transactions associated with the block;
\item \code{net_usage} (integer): total bandwidth consumed by the block's transactions;
\item \code{energy_usage} (integer): total energy used by the block's transactions.
}
}
\description{
Retrives attributes of the blocks produced within a user-specified time range
}
\examples{
r <- get_blocks_for_time_range(
  min_timestamp = "1551715200000",
  max_timestamp = "1551715210000"
)
print(r)
}
