% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chain_statistics.R
\name{get_chain_statistics}
\alias{get_chain_statistics}
\title{Get historical metrics of the TRON blockchain}
\usage{
get_chain_statistics(
  days = 14,
  include_current_date = FALSE,
  max_attempts = 3L
)
}
\arguments{
\item{days}{(double): number of days to look back (defaults to 14 days).}

\item{include_current_date}{(boolean): whether the current date's statistics
(as of the query time) should be returned. Defaults to \code{FALSE}.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{date} (Date): date in the \code{YYYY-MM-DD} format;
\item \code{avg_block_time} (double): average time covered by a block (in seconds);
\item \code{avg_block_size} (double): average block size (in bytes);
\item \code{total_blocks} (integer): cumulative number of blocks on the chain;
\item \code{new_blocks} (integer): number of newly generated blocks;
\item \code{chain_size} (double): cumulative size of the chain (in bytes);
\item \code{total_addresses} (integer): cumuative number of address on the chain;
\item \code{active_addresses} (integer): number of addresses that were active on \code{date};
\item \code{new_addresses} (integer): number of newly created addresses;
\item \code{addresses_with_trx} (integer): number of Tronix (TRX)-holding addresses;
\item \code{total_trc10_tokens} (integer): cumulative number of TRC-10 assets on the
chain;
\item \code{total_trc20_tokens} (integer): cumulative number of TRC-20 assets on the
chain;
\item \code{new_trc10_tokens}(integer): number of newly created TRC-10 assets;
\item \code{new_trc20_tokens} (integer): number of newly created TRC-20 assets;
\item \code{new_tx} (integer): number of new transaction on the chain;
\item \code{trx_transfer_tx} (integer): number of TRX transfer transactions;
\item \code{trc10_transfer_tx} (integer): number of TRC-10 transfer transactions;
\item \code{freeze_tx} (integer): number of TRX freezing transactions;
\item \code{vote_tx} (integer): number of vote transactions;
\item \code{shielded_tx} (integer): number of shielded transactions;
\item \code{other_tx} (integer): number of other transactions;
\item \code{contract_triggers} (integer): cumulative number of smart contract triggers;
\item \code{energy_usage} (double): amount of energy consumed;
\item \code{net_usage} (double): amount of bandwidth consumed.
}
}
\description{
Retrieves historical values of the blockhain statistics
}
\examples{
r <- get_chain_statistics(days = 7)
print(r)
}
