% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_trx_market_data.R
\name{get_current_trx_market_data}
\alias{get_current_trx_market_data}
\title{Get TRX market data}
\usage{
get_current_trx_market_data(vs_currencies = c("usd", "eur"), max_attempts = 3L)
}
\arguments{
\item{vs_currencies}{(character): a vector with
names of the base currencies to benchmark TRX against, e.g.
\code{c("usd", "eur", "btc")}. An up-to-date list of supported currencies
(both fiat and cryptocurrencies) can be retrieved with the
\code{\link[=get_supported_coingecko_currencies]{get_supported_coingecko_currencies()}} function. If \code{vs_currencies}
contains at least one unsupported currency, the call to this function
will fail with the respective error message.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with as many rows as the length of \code{vs_currencies}. Please
refer to the \href{https://www.coingecko.com/en/methodology}{"Methodology"}
page on the CoinGecko website for definitions of variables in this
tibble.
}
\description{
Retrieves the current market data for TRX (price, market cap, high/low, etc.)
}
\examples{
r <- get_current_trx_market_data(vs_currencies = c("usd", "eur"))
print(r)
}
