% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_trx_price.R
\name{get_current_trx_price}
\alias{get_current_trx_price}
\title{Get Tronix price}
\usage{
get_current_trx_price(
  vs_currencies = c("usd"),
  include_market_cap = FALSE,
  include_24h_vol = FALSE,
  include_24h_change = FALSE,
  max_attempts = 3L
)
}
\arguments{
\item{vs_currencies}{(character): a vector with names of the currencies
to express the TRX price in, e.g. \code{c("usd", "eur", "btc")}. An up-to-date
list of supported currencies (both fiat and cryptocurrencies) can
be found at \url{https://api.coingecko.com/api/v3/simple/supported_vs_currencies}}

\item{include_market_cap}{(boolean, defaults to \code{FALSE}): whether to return
the market cap information.}

\item{include_24h_vol}{(boolean, defaults to \code{FALSE}): whether to return
the trading volume for the last 24 hours.}

\item{include_24h_change}{(boolean, defaults to \code{FALSE}): whether to return
the price percentage change compared to 24 hours ago.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with as many rows as the length of \code{vs_currencies} and at
at least 3 columns: \code{trx_price} (double), \code{vs_currency} (character),
and \code{last_updated_at} (POSIXct, UTC timezone). If the boolean arguments
of this function are all \code{TRUE}, then the resultant tibble will
include additional three columns: \code{market_cap} (double),
\code{vol_24h} (double) and \code{price_percent_change_24h} (double).
}
\description{
Returns the current TRX price and market data against other currencies
}
\details{
This function is based on the public
CoinGecko API, which has a limit of 100 calls per minute. Please
keep this limit in mind when developing your code.
}
\examples{
r <- get_current_trx_price(vs_currencies = c("usd", "eur"))
print(r)
}
