% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_by_tx_id.R
\name{get_events_by_tx_id}
\alias{get_events_by_tx_id}
\title{Get events by transaction ID}
\usage{
get_events_by_tx_id(
  tx_id,
  only_confirmed = NULL,
  only_unconfirmed = NULL,
  max_attempts = 3L
)
}
\arguments{
\item{tx_id}{(character): transaction ID.}

\item{only_confirmed}{(boolean or \code{NULL}): if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for confirmed
transactions are returned.}

\item{only_unconfirmed}{(boolean or \code{NULL}): if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for
unconfirmed transactions are returned. Cannot be used
simultanously with the \code{only_confirmed} argument.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble where each row corresponds to an event associated
with the account of interest. This tibble contains the following
columns:
\itemize{
\item \code{tx_id} (character): transaction ID;
\item \code{block_number} (character);
\item \code{block_timestamp} (POSIXct, UTC timezone);
\item \code{contract_address} (character): adress of the smart contract that implemented
the event;
\item \code{event_name} (character): possible values of this column are contract-
and event-specific;
\item \code{event_data} (list): each element of this list contains a named list with
additional attributes of the event.
}

If no events are found for the specified combination of query
parameters, nothing (\code{NULL}) is returned, with a console message
\code{"No data found"}.
}
\description{
Retrieves events associated with a transaction
}
\details{
The exact content of \code{event_data} in the tibble returned by this
function will be contract- and event-specific. Thus, these data are
returned in the form of a named R list as-is, i.e. without any
processing. Users are advised to develop their own, task-specific, logic
to process event attributes.
}
\examples{
tx_id <- "270e992bf22a271008032ec09a51616ed963b74f4d808b0fd74fc82341a81391"
r <- get_events_by_tx_id(tx_id)
print(r)
}
