% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_of_latest_block.R
\name{get_events_of_latest_block}
\alias{get_events_of_latest_block}
\title{Get events of latest block}
\usage{
get_events_of_latest_block(only_confirmed = FALSE, max_attempts = 3L)
}
\arguments{
\item{only_confirmed}{(boolean or \code{NULL}) - if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for confirmed
transactions are returned.}

\item{max_attempts}{(integer, positive) - specifies the
maximum number of additional attempts to call the API if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to 3.}
}
\value{

}
\description{
Retrieves events that took place in the most recent block
}
\examples{
r <- get_events_of_latest_block()
print(r)

}
