% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_block.R
\name{get_latest_block}
\alias{get_latest_block}
\title{Get the latest block}
\usage{
get_latest_block(max_attempts = 3L)
}
\arguments{
\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble with the following columns:
\itemize{
\item \code{block_number} (integer): number of the latest block;
\item \code{timestamp} (POSIXct): date and time when the block was created;
\item \code{hash} (character): hash of the block;
\item \code{parent_hash} (character): hash of the parent block;
\item \code{tx_trie_root} (character): hash of the
\href{https://en.wikipedia.org/wiki/Trie}{trie} root of the block creation
transaction;
\item \code{confirmed} (boolean): an indicator of whether this block has been
confirmed;
\item \code{size} (integer): size of the block (in bytes);
\item \code{witness_address} (character): address of the block's creator;
\item \code{tx_count} (integer): number of transactions associated with this block;
\item \code{tx} (list): a list with one element that contains a tibble with basic
info on the transactions associated with this block: \code{tx_id} (character) -
transaction ID, \code{from_address} (character) - address of the account that
initiated the transaction, and \code{to_address} (character) - address of the
recieving account. All addresses are presented in the \code{base58check} format.
}
}
\description{
Retrieves information on the latest block and its transactions
}
\details{
Additional details on the block producing process can be found
in the \href{https://tronprotocol.github.io/documentation-en/introduction/dpos/#dpos}{official documentation}.
}
\examples{
r <- get_latest_block()
print(r)

}
