% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supported_coingecko_currencies.R
\name{get_supported_coingecko_currencies}
\alias{get_supported_coingecko_currencies}
\title{Get CoinGecko currencies}
\usage{
get_supported_coingecko_currencies(max_attempts = 3L)
}
\arguments{
\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
Character vector with abbreviated names of the currencies.
}
\description{
Retrieves a list of base currencies currently supported by the CoinGecko API
}
\examples{
r <- get_supported_coingecko_currencies()
print(r)
}
