% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trc10_transfers.R
\name{get_trc10_transfers}
\alias{get_trc10_transfers}
\title{Get TRC-10 transfers}
\usage{
get_trc10_transfers(
  owner_address = NULL,
  min_timestamp,
  max_timestamp,
  max_attempts = 3L
)
}
\arguments{
\item{owner_address}{(character): address of the
token issuer, either in \code{base58check} or \code{hex} format.}

\item{min_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
beginning of the period of interest (inclusive). Defaults to 0.}

\item{max_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
end of the period of interest (inclusive).}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
Tibble with the following colums (see \code{\link[=get_tx_info_by_id]{get_tx_info_by_id()}} and
\code{\link[=get_trc10_token_description]{get_trc10_token_description()}} for their definitions):
\itemize{
\item \code{tx_id} (character);
\item \code{block_number} (character);
\item \code{timestamp} (POSIXct, UTC timezone);
\item \code{from_address} (character);
\item \code{to_address} (character);
\item \code{is_contract_from_address} (boolean);
\item \code{is_contract_to_address} (boolean);
\item \code{contract_result} (character);
\item \code{confirmed} (boolean);
\item \code{amount} (double);
\item \code{token_id} (character);
\item \code{token_name} (character);
\item \code{token_abbr} (character).
}
}
\description{
Returns TRC-10 transfer transactions for a user-specified time range
}
\details{
If \code{owner_address = NULL}, all TRC-10 asset transfers will be
returned, including Tronix (TRX) (which is technically a TRC-10 token).
However, if \code{owner_address} is specified, only results for the
corresponding token will be returned. This argument accepts only a single
address at a time.

The number of transfers that take place on the TRON blockchain can be
very large, and thus users are advised to choose \code{min_timestamp} and
\code{max_timestamp} wisely. If the requested time range is too large, the
maximum number of transactions returned by the underlying Tronscan API
will be \emph{capped} at 10000, and the processing time may become
prohibitively long. Chunking the time range of interest into
smaller periods can help to avoid gaps in data in such cases. However,
users would have to implement their own logic for that.
}
\examples{
# Results contain all TRC-10 transfers that took place
# within the specified time range (including TRX):
r1 <- get_trc10_transfers(
  min_timestamp = "1577837400000",
  max_timestamp = "1577837430000"
)
print(r1)

# Results contain transfers of a specific token:
r2 <- r <- get_trc10_transfers(
  owner_address = "THLLMnsEKEci5e5dJHnW28QQU8AujGhSoK",
  min_timestamp = "1577837400000",
  max_timestamp = "1577837430000"
)
print(r2)
}
