% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trc20_token_description.R
\name{get_trc20_token_description}
\alias{get_trc20_token_description}
\title{Get description of a TRC-20 token}
\usage{
get_trc20_token_description(
  contract_address,
  detailed_info = FALSE,
  max_attempts = 3L
)
}
\arguments{
\item{contract_address}{(character): contract address of
a TRC-20 token of interest, either in \code{base58check} or
\code{hex} format.}

\item{detailed_info}{(boolean): if \code{FALSE} (default),
only basic information about the token is returned.
If \code{TRUE}, an extended information is returned.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble, whose content depends on the \code{detailed_info} argument. If
\code{detailed_info = FALSE} (default), the tibble will have the following 5
columns:
\itemize{
\item \code{token_name} (character): commont name of the token;
\item \code{token_abbr} (character): abbreviated name of the token ("symbol");
\item \code{token_contract_address} (character): address of the token's contract, in
\code{base58check} format;
\item \code{token_owner_address} (character): address of the token's creator, in
\code{base58check} format;
\item \code{precision} (integer): number of digits in the decimal part of the
token's amount (see \code{\link[=apply_decimal]{apply_decimal()}} for details).
}

If \code{detailed_info = TRUE}, the returned tibble will have the same 5 columns
as above, and the following additional columns:
\itemize{
\item \code{request_time} (POSIXct): date and time when the query was made;
\item \code{vip} (boolean): indicator of whether this token is treated as a VIP asset
on the blockchain;
\item \code{description} (character): a free-text field describing the token;
\item \code{date_created} (POSIXct): date and time of the token's creation;
\item \code{url} (character): URL of the token's project;
\item \code{total_supply} (double): total supply of the token;
\item \code{number_of_holders} (integer): current number of accounts that hold this
token;
\item \code{total_tx} (integer): current cumulative number of transactions that
involved this token;
\item \code{price_in_trx} (double): current price of this token expressed in TRX.
}

If no description for the requested token can be found, the function will
return no data (\code{NULL}), with the respective console message.
}
\description{
Returns various bits of information about a TRC-20 token
}
\details{
TRC-20 are tokens issued by smart contracts. See
\href{https://developers.tron.network/docs/trc20}{official documentation}
for details.
}
\examples{
r1 <- get_trc20_token_description(
  contract_address = "TLa2f6VPqDgRE67v1736s7bJ8Ray5wYjU7",
  detailed_info = TRUE
)
r2 <- get_trc20_token_description(
  contract_address = "TLa2f6VPqDgRE67v1736s7bJ8Ray5wYjU7",
  detailed_info = FALSE
)
print(r1)
print(r2)
}
