% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trc20_tx_info_by_account_address.R
\name{get_trc20_tx_info_by_account_address}
\alias{get_trc20_tx_info_by_account_address}
\title{Get TRC20 transactions for a given account}
\usage{
get_trc20_tx_info_by_account_address(
  address,
  only_confirmed = NULL,
  only_unconfirmed = NULL,
  only_to = FALSE,
  only_from = FALSE,
  min_timestamp = 0,
  max_timestamp = NULL,
  contract_address = NULL,
  limit = 200L,
  max_attempts = 3L
)
}
\arguments{
\item{address}{(character) - address of the account of interest, in
\code{base58} (starts with \code{T}) or \code{hex} (starts with \code{41}) format.}

\item{only_confirmed}{(boolean or \code{NULL}) - if \code{NULL} (default) or \code{FALSE},
both confirmed and unconfirmed transactions are returned. If \code{TRUE},
only confirmed transactions are returned. Cannot be used simultanously
with the \code{only_unconfirmed} argument (see next).}

\item{only_unconfirmed}{(boolean or \code{NULL}) - if \code{NULL} (default) or \code{FALSE},
both confirmed and unconfirmed transactions are returned. If \code{TRUE},
only unconfirmed transactions are returned. Cannot be used simultanously
with the \code{only_confirmed} argument.}

\item{only_to}{(boolean, defautls to \code{FALSE}) - if \code{TRUE}, only
inbound transactions are returned.}

\item{only_from}{(boolean, defautls to \code{FALSE}) - if \code{TRUE}, only
outbound transactions are returned.}

\item{min_timestamp}{(numeric or character) - a Unix timestamp
(\emph{including milliseconds}), which defines the beginning of the
period to retrieve the transactions from. Defaults to 0.}

\item{max_timestamp}{(numeric or character) - a Unix timestamp
(\emph{including milliseconds}), which defines the end of the
period to retrieve the transactions from.}

\item{contract_address}{(character) - address of the TRC20 token's
contract, in \code{base58} or \code{hex} format.}

\item{limit}{(integer) - number of transactions per page. Defaults to 200.
Maximum accepted value is 200 (higher values will be ignored).}

\item{max_attempts}{(integer, positive) - specifies the maximum
number of additional attempts to call the API if the first attempt fails
(i.e. its call status is different from \code{200}). Additional attempts are
implemented with an exponential backoff. Defaults to 3.}
}
\value{
A tibble where each row corresponds to one transaction.
This tibble contains the following columns:
\itemize{
\item \code{tx_id} (character) - transation ID;
\item \code{tx_type} (character) - transation type (e.g., \code{"Transfer"});
\item \code{block_timestamp} (POSIXct);
\item \code{from_address} (character);
\item \code{to_address} (character);
\item \code{trc20_symbol} (character) - an abbreviated name of the TRC20 token;
\item \code{trc20_name} (character) - a common name of the TRC20 token;
\item \code{trc20_contract_address} (character);
\item \code{precision} (integer) - precision of the \code{amount} values;
\item \code{amount} (character) - transaction amount of the TRC20 token.
}

If no TRC20 are found for the specified combnations of query parameters,
nothing (NULL) is returned, with a console message
\code{"No TRC20 transactions found for this account within this time range"}.
}
\description{
Returns various bits of information about the TRC20 transactions
that went through a given account
}
\details{
Some accounts may have a very high load of transactions going
through them. Users are, therefore, advised to choose \code{min_timestamp} and
\code{max_timestamp} wisely as the TronGrid API will not return more than
10000 transactions in one request. If the number of transactions
exceeds this limit, the the request will fail with status \code{404} and an
error message
\code{"Exceeds the maximum limit, please change your query time range"}). If
data are to be retrieved for a large time range, users are advised
to implement their own logic for querying the network (e.g.,
splitting that range into smaller chunks and then combining the results).
}
\examples{
tx_df <- get_tx_info_by_account_address(
                         address = "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX",
                         min_timestamp = "1604188800000",
                         max_timestamp = "1604189100000",
                         limit = 20L)
print(tx_df)

}
