% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trx_market_data_for_date.R
\name{get_trx_market_data_for_date}
\alias{get_trx_market_data_for_date}
\title{Get TRX market data for a date}
\usage{
get_trx_market_data_for_date(
  date,
  vs_currencies = c("usd", "eur"),
  max_attempts = 3L
)
}
\arguments{
\item{date}{(character, Date or POSIXct): date to retrieve the data for.
Expected format: \verb{\%Y-\%m-\%d}.}

\item{vs_currencies}{(character): a vector with
names of the base currencies to benchmark TRX against, e.g.
\code{c("usd", "eur", "btc")}. An up-to-date list of supported currencies
(both fiat and cryptocurrencies) can be retrieved with the
\code{\link[=get_supported_coingecko_currencies]{get_supported_coingecko_currencies()}} function. If \code{vs_currencies}
contains at least one unsupported currency, the call will fail with the
respective error message.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with as many rows as the length of \code{vs_currencies} and
the following columns:
\itemize{
\item \code{date} (Date): same as \code{date};
\item \code{vs_currency} (character): same as \code{vs_currencies};
\item \code{market_cap} (double): market cap as of \code{date};
\item \code{total_trading_vol} (double): total trading volume on \code{date};
\item \code{price} (double): average price on \code{date}.
}
}
\description{
Retrieves TRX market data for a specific date (price, market cap, high/low, etc.)
}
\examples{
r <- get_trx_market_data_for_date(date = Sys.Date() - 1)
print(r)
}
