% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trx_market_data_for_date_range.R
\name{get_trx_market_data_for_date_range}
\alias{get_trx_market_data_for_date_range}
\title{Get historical TRX market data}
\usage{
get_trx_market_data_for_date_range(
  vs_currency = "usd",
  from_date,
  to_date,
  max_attempts = 3L
)
}
\arguments{
\item{vs_currency}{(character): name of the base currency to benchmark TRX
against (\code{usd} by default). An up-to-date list of supported currencies
(both fiat and cryptocurrencies) can be retrieved with the
\code{\link[=get_supported_coingecko_currencies]{get_supported_coingecko_currencies()}} function. If an unsupported
\code{vs_currency} is requested, the call will fail with the respective error
message.}

\item{from_date}{(POSIXct): start of the time range to retrieve the data for.
Expected format: \verb{\%Y-\%m-\%d \%H:\%M:\%S}. The minimal acceptable \code{from_date}
is \verb{2017-11-09 00:00:00} as no data are available for earlier dates.
Attempts to retrieve data for earlier dates will fail with the
corresponding error.}

\item{to_date}{(POSIXct): end of the time range to retrieve the data for
(inclusive). Expected format: \verb{\%Y-\%m-\%d \%H:\%M:\%S}. Attempts to request a
future \code{to_date} for which no history exists yet will fail with the
corresponding error.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{datetime} (POSIXct): timestamp;
\item \code{vs_currency} (character): same as the argument \code{vs_currency};
\item \code{price} (double): TRX price, as of \code{datetime};
\item \code{total_trading_vol} (double): a 24 h rolling-window trading volume, as
of \code{datetime};
\item \code{market_cap} (double): TRX market cap, as of \code{datetime}.
}
}
\description{
Retrieves TRX market data for a range of historical dates
}
\details{
Minutely data will typically be used for periods of <1 day,
hourly data for periods between 1 day and 90 days, and daily data for
periods above 90 days. Please note that these data granularity intervals
are only approximate and a substantial variation may be observed.
}
\examples{
from <- as.POSIXct("2021-01-01 10:00:10", tz = "UTC")
to <- as.POSIXct("2021-01-01 20:45:00", tz = "UTC")
r <- get_trx_market_data_for_date_range(
  vs_currency = "eur",
  from_date = from,
  to_date = to
)
print(r)
}
