% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trx_market_data_for_last_n_days.R
\name{get_trx_market_data_for_last_n_days}
\alias{get_trx_market_data_for_last_n_days}
\title{Get historical TRX market data}
\usage{
get_trx_market_data_for_last_n_days(
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3L
)
}
\arguments{
\item{vs_currency}{(character): name of the base currency to benchmark TRX
against (\code{usd} by default). An up-to-date list of supported currencies
(both fiat and cryptocurrencies) can be retrieved with the
\code{\link[=get_supported_coingecko_currencies]{get_supported_coingecko_currencies()}} function. If an unsupported
\code{vs_currency} is requested, the call will fail with the respective error
message.}

\item{days}{(numeric or \code{"max"}): number of days to look back. If the
requested number of \code{days} covers dates before \code{2017-11-09}, the retrived
data will be clipped at \code{2017-11-09} (the beginning of history for TRX).
If \code{days = "max"}, the entire available history will be retrieved.
Depending on the value of \code{days}, the time interval used to present the
data will differ - see "Details".}

\item{interval}{(character or \code{NULL}): time interval to present the data.
The only currently supported interval is \code{daily}. Defaults to \code{NULL}.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{datetime} (POSIXct): timestamp;
\item \code{vs_currency} (character): same as the argument \code{vs_currency};
\item \code{price} (double): TRX price, as of \code{datetime};
\item \code{total_trading_vol} (double): a 24 h rolling-window trading volume, as
of \code{datetime};
\item \code{market_cap} (double): TRX market cap, as of \code{datetime}.
}
}
\description{
Retrieves TRX market data for the last \code{n} days
}
\details{
If \code{days = 1} and \code{interval = NULL}, the data will be presented for
every few minutes (typically 3-8 minutes). If \code{days} is between 2 and 90
(inclusive) and \code{interval = NULL}, an (approximately) hourly time step will
be used. Daily data are used for \code{days} above 90. If \code{interval = "daily"},
daily data will be used irrespective of the value of \code{days}.
}
\examples{
r <- get_trx_market_data_for_last_n_days(
  vs_currency = "gbp",
  days = 30
)
print(r)
}
