% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trx_market_data_for_time_range.R
\name{get_trx_market_data_for_time_range}
\alias{get_trx_market_data_for_time_range}
\title{Get historical TRX market data}
\usage{
get_trx_market_data_for_time_range(
  vs_currency = "usd",
  min_timestamp,
  max_timestamp,
  max_attempts = 3L
)
}
\arguments{
\item{vs_currency}{(character): name of the base currency to benchmark TRX
against (\code{usd} by default). An up-to-date list of supported currencies
(both fiat and cryptocurrencies) can be retrieved with the
\code{\link[=get_supported_coingecko_currencies]{get_supported_coingecko_currencies()}} function. If an unsupported
\code{vs_currency} is requested, the call will fail with the respective error
message.}

\item{min_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
beginning of the period of interest (inclusive). Defaults to 0.}

\item{max_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
end of the period of interest (inclusive).}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{timestamp} (POSIXct);
\item \code{vs_currency} (character): same as the argument \code{vs_currency};
\item \code{price} (double): TRX price, as of \code{datetime};
\item \code{total_trading_vol} (double): a 24 h rolling-window trading volume, as
of \code{timestampt};
\item \code{market_cap} (double): TRX market cap, as of \code{timestamp}.
}
}
\description{
Retrieves TRX market data for a range of historical dates
}
\details{
This function returns hourly data for periods of up to 90 days,
and daily data for periods above 90 days.

The minimal acceptable \code{min_timestamp} is \code{"1510185600000"} (which
corresponds to \verb{2017-11-09 00:00:00}) as no data are available for
earlier dates. Attempts to retrieve data for earlier dates will fail
with the corresponding error message.

Attempts to request a future \code{max_timestamp} for which no history exists
yet will also fail with the corresponding error message.
}
\examples{
r <- get_trx_market_data_for_time_range(
  vs_currency = "eur",
  min_timestamp = "1609495210000",
  max_timestamp = "1609533900000"
)
print(r)
}
