% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trx_ohlc_data_for_last_n_days.R
\name{get_trx_ohlc_data_for_last_n_days}
\alias{get_trx_ohlc_data_for_last_n_days}
\title{Get historical TRX price data}
\usage{
get_trx_ohlc_data_for_last_n_days(vs_currency = "usd", days, max_attempts = 3L)
}
\arguments{
\item{vs_currency}{(character): name of the base currency to benchmark TRX
against (\code{usd} by default). An up-to-date list of supported currencies
(both fiat and cryptocurrencies) can be retrieved with the
\code{\link[=get_supported_coingecko_currencies]{get_supported_coingecko_currencies()}} function. If an unsupported
\code{vs_currency} is requested, the call will fail with the respective error
message.}

\item{days}{(numeric or \code{"max"}): number of days to look back. If
\code{days = "max"}, the entire available history will be retrieved. Depending
on the value of \code{days}, the time interval used to present the data will
differ - see "Details".}

\item{max_attempts}{function_params("max_attempts")}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{datetime} (POSIXct): timestamp;
\item \code{vs_currency} (character): same as the argument \code{vs_currency};
\item \code{price_open} (double): TRX price in the beginning of a time iterval;
\item \code{price_high} (double): highest TRX price observed within a time interval;
\item \code{price_low} (double): lowest TRX price observed within a time interval;
\item \code{price_close} (double): TRX price in the end of a time interval.
}
}
\description{
Retrieves TRX open-high-low-close price data for the last \code{n} days
}
\details{
Granularity of the retrieved data
(i.e. \href{https://en.wikipedia.org/wiki/Open-high-low-close_chart}{candle}'s
body) depends on the number of requested \code{days} as follows:
\itemize{
\item 1 - 2 days: 30 minutes
\item 3 - 30 days: 4 hours
\item 31 and before: 4 days
}
}
\examples{
r <- get_trx_ohlc_data_for_last_n_days(days = 7)
print(r)
}
