% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_info_by_id.R
\name{get_tx_info_by_id}
\alias{get_tx_info_by_id}
\title{Get transaction attributes}
\usage{
get_tx_info_by_id(tx_id, add_contract_data = TRUE, max_attempts = 3L)
}
\arguments{
\item{tx_id}{(character): transaction ID.}

\item{add_contract_data}{(boolean): if \code{TRUE} (default),
adds column \code{contract_data} to the resultant tibble (see Value).}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble with one row and the following columns:
\itemize{
\item \code{request_time} (POSIXct): date and time when the request was made;
\item \code{tx_id} (character): transation ID;
\item \code{block_number} (character);
\item \code{timestamp} (POSIXct, UTC timezone): block's time stamp;
\item \code{contract_result} (character): result of calling the contract
(e.g., \code{SUCCESS}, \code{REVERT}, etc.);
\item \code{confirmed} (boolean): a flag indicating whether this transaction has been
confirmed;
\item \code{confirmation_count} (integer): number of accounts that confirmed;
\item \code{sr_confirm_list} (list): contains a tibble with the following
columns:
\itemize{
\item \code{address} (character): \code{base58check}-formatted address of the
Super Representative's account that confirmed this transaction;
\item \code{name} (character): name of the Super Representative's account;
\item \code{block_number} (character): block that the confirmation transaction
belongs to.
}
\item \code{contract_type} (character): type of the system contract call (see
\href{https://tronprotocol.github.io/documentation-en/mechanism-algorithm/system-contracts/}{official documentation}
for details);
\item \code{from_address} (character): \code{base58check}-formatted address of the account
that initiated this transaction (also known as the "owner address");
\item \code{to_address} (character): \code{base58check}-formatted address of the receiving
account;
\item \code{is_contract_from_address_account} (boolean): flag indicating whether the
\code{from_address} is a contract account;
\item \code{is_contract_to_address} (boolean): flag indicating whether the
\code{to_address} is a contract account;
\item \code{costs} (list): contains a tibble with the following columns:
\itemize{
\item \code{net_fee} (integer);
\item \code{net_fee_cost} (integer);
\item \code{energy_usage} (integer);
\item \code{energy_fee_cost} (integer);
\item \code{energy_fee} (integer);
\item \code{energy_usage_total} (integer);
\item \code{origin_energy_usage} (integer);
\item \code{net_usage} (integer).
}
\item \code{tx_transfer} (double): amount of TRX transferred from \code{from_address} to
\code{to_address} as part of this transaction;
\item \code{trc10_transafer} (list): contains a tibble with information on the
TRC-10 tokens (other than TRX) transferred as part of this transaction
(or \code{NA} if no such transfer happened). This tibble contains the following
columns:
\itemize{
\item \code{token_id} (character): ID of the TRC-10 token transferred;
\item \code{token_name} (character): common name of the token;
\item \code{token_abbr} (character): abbreviated name of the token;
\item \code{vip} (boolean): flag indicating whether this token is treated as a
VIP asset on the TRON blockchain;
\item \code{amount} (double): amount of the token transferred.
}
\item \code{trc20_transfer} (list): contains a tibble with iformation on the TRC-20
tokens transferred as part of this transaction (or \code{NA} if no such transfer
happened). This tibble contains the following columns:
\itemize{
\item \code{token_name} (character): common name of the token;
\item \code{token_abbr} (character): abbreviated name of the token;
\item \code{token_contract} (character): \code{base58check}-formatted address of the
token's contract;
\item \code{from_address} (character): \code{base58check}-formatted address of the
sending account;
\item \code{to_address} (character): \code{base58check}-formatted address of the
receiving account;
\item \code{is_contract_from_address_account} (boolean): flag indicating whether
the \code{from_address} is a contract account;
\item \code{is_contract_to_address} (boolean): flag indicating whether the
\code{to_address} is a contract account;
\item \code{vip} (boolean): flag indicating whether this token is treated as a
VIP asset on the TRON blockchain;
\item \code{amount} (double): amount of the token transferred.
}
\item \code{internal_tx} (list): contains a tibble with iformation on the internal
transactions triggered by this transaction (or \code{NA} if no internal
transactions occurred). This tibble contains the following columns:
\itemize{
\item \code{internal_tx_id} (character): ID of the internal transaction;
\item \code{from_address} (character): \code{base58check}-formatted address that
initiated this internal transaction (also known as "owner address");
\item \code{to_address} (character): \code{base58check}-formatted address of the
receiving account;
\item \code{is_contract_from_address_account} (boolean): flag indicating whether
the \code{from_address} is a contract account;
\item \code{is_contract_to_address} (boolean): flag indicating whether the
\code{to_address} is a contract account;
\item \code{confirmed} (boolean): a flag indicating whether this transaction has
been confirmed;
\item \code{rejected} (boolean): a flag indicating whether this transaction has
been rejected for some reason;
\item \code{token_id} (character): ID of the token transferred as part of this
internal transaction;
\item \code{token_name} (character): common name of the token;
\item \code{token_abbr} (character): abbreviated name of the token;
\item \code{vip} (boolean): flag indicating whether this token is treated as a
VIP asset on the TRON blockchain;
\item \code{amount} (double): amount of the token transferred.
}
\item \code{info} (list): contains a list with additional attributes of this
transaction (or \code{NA} if no such information is available);
\item \code{contract_data} (list): contains a list with raw data generated by the
contract that performed this transaction. This column is only added if the
\code{add_contract_data} argument is \code{TRUE}. The actual content of the list with
raw data will depend on the transaction's nature.
}
}
\description{
Returns attributes of a transaction based on its ID
}
\details{
All token amounts in the resultant tibble (specifically, in
columns \code{trx_transfer}, \code{trc10_transfer}, \code{trc20_transfer}, and
\code{internal_tx}) are expressed using the whole number and the decimal
parts. However, if the user requests raw contract data
(\code{add_contract_data = TRUE}), the returned raw data will show token
amounts "as is", i.e. expressed using the machine-level precision.
See \code{\link[=apply_decimal]{apply_decimal()}} for details.
}
\examples{
id <- "dca447279bc2fd3c10325e442746f9a42938e25bac33bc277b3e7720027aaaf2"
tx <- get_tx_info_by_id(id)
print(tx)
}
