% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_address.R
\name{is_address}
\alias{is_address}
\title{Check if a given address is valid}
\usage{
is_address(address)
}
\arguments{
\item{address}{(character) An address to be validated.}
}
\value{
\code{TRUE} or \code{FALSE}, depending on the result.
}
\description{
Checks if a given address is a valid \code{base58check}- or \code{hex}-formatted address
}
\examples{
is_address("41357a7401a0f0c2d4a44a1881a0c622f15d986291") # TRUE
is_address("TEqyWRKCzREYC2bK2fc3j7pp8XjAa6tJK1") # TRUE
is_address("abc") # FALSE

}
