% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_assets_on_chain.R
\name{list_all_assets_on_chain}
\alias{list_all_assets_on_chain}
\title{List all TRC-10 assets}
\usage{
list_all_assets_on_chain(
  order_by = "total_supply",
  direction = "desc",
  max_attempts = 3L
)
}
\arguments{
\item{order_by}{(character): specifies the variable to
order by (see details).}

\item{direction}{(character): specifies the direction of
ordering the results - descending (\code{desc}) or ascending (\code{asc}).}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{request_time} (POSIXct): date and time when the API request was made;
\item \code{asset_id} (character): asset ID, presented as a set of numbers
(e.g. \code{1002762});
\item \code{owner_address} (character): address of the asset issuer, in \code{base58check}
format;
\item \code{abbr} (character): abbreviated name of the asset;
\item \code{asset_name} (character): full name of the asset;
\item \code{precision} (integer): precision used to present the asset's balance
(e.g., if it's 6, then one needs to divide the returned balance by 1
million to obtain the actual balance for that asset).
\item \code{description} (character): a free-text field describing the asset;
\item \code{url} (character): URL of the token's project;
\item \code{total_supply} (character): total issued amount tokens;
\item \code{num} (character): amount of the asset tokens that one can buy
with \code{trx_num} TRX tokens (see next point);
\item \code{trx_num} (character): amount of TRX tokens that is required to buy \code{num}
tokens of the asset (thus, \code{num / num_trx} is the asset's price during
its ICO);
\item \code{ico_start_time} (POSIXct, UTC timezone): date and time of the asset's ICO
start;
\item \code{ico_end_time} (POSIXct, UTC timezone): date and time of the asset's ICO
end.
\item \code{vote_score} (integer): vote score.
}
}
\description{
Returns a list of all TRC-10 assets currently available on the chain
}
\details{
TRC-10 are tokens issued by the system contract (as opposed to
TRC-20, which are issued by smart contracts). See
\href{https://developers.tron.network/docs/trc10}{official documentation}
for details.

Argument \code{order_by} expects one of the following values: \code{total_supply},
\code{ico_start_time}, \code{ico_end_time}, \code{id}.
}
\examples{
r <- list_all_assets_on_chain(
  order_by = "total_supply",
  direction = "desc"
)
print(r)
}
