% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_top_trx_holders.R
\name{list_top_trx_holders}
\alias{list_top_trx_holders}
\title{List top TRX holders}
\usage{
list_top_trx_holders(n = 20, max_attempts = 3L)
}
\arguments{
\item{n}{(double): number of top accounts to retrieve.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble with \code{n} rows and the following columns:
\itemize{
\item \code{request_time} (POSIXct): date and time when the request was made;
\item \code{address} (character): account address (in \code{base58check} format);
\item \code{trx_balance} (double): TRX balance;
\item \code{total_tx} (integer): total number of transactions associated with
the respective \code{account};
\item \code{tron_power} (double): amount of TRX frozen
(see \href{https://tronprotocol.github.io/documentation-en/introduction/overview/#2-srs-and-committee}{official documentation}
for details).
}
}
\description{
Returns a list of accounts with the largest TRX balances
}
\examples{
r <- list_top_trx_holders(10)
print(r)
}
