% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_account_balance_info.R
\name{parse_account_balance_info}
\alias{parse_account_balance_info}
\title{Parse account balance attributes}
\usage{
parse_account_balance_info(info)
}
\arguments{
\item{info}{A non-empty, named list returned as a result of calling the
\code{api/account} endpoint of the Tronscan API. This list is also
supposed to be appended with a \code{request_time} element that contains
a POSIXct values correspinding to the time when the API call was made.}
}
\value{
A nested tibble with parsed attributes related to the account
balance. See \code{\link[=get_account_balance]{get_account_balance()}} for details.
}
\description{
Converts a list with account balance attributes into a nested tibble
}
\keyword{internal}
