% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tx_info.R
\name{parse_tx_info}
\alias{parse_tx_info}
\title{Parse transaction attributes}
\usage{
parse_tx_info(info)
}
\arguments{
\item{info}{(list): a non-empty, named list returned as a result of calling the
\href{https://github.com/tronscan/tronscan-frontend/blob/dev2019/document/api.md}{\verb{/api/transaction-info}}
endpoint of the Tronscan API.}
}
\value{
A nested tibble with one row and several columns that contain
transaction attributes. See \code{\link[=get_tx_info_by_id]{get_tx_info_by_id()}} for details.
}
\description{
Converts a list with transaction attributes into a nested tibble
}
\keyword{internal}
