% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_paginated_tronscan_query.R
\name{run_paginated_tronscan_query}
\alias{run_paginated_tronscan_query}
\title{Run paginated Tronscan queries}
\usage{
run_paginated_tronscan_query(base_url, path, params, max_attempts = 3L)
}
\arguments{
\item{base_url}{(character): host URL.}

\item{path}{(character): a vector of character values that define the
API's endpoint. Order of these values matters - it must be same as in the
endpoint's definition.}

\item{params}{(list): a named list with query parameters.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A list with the retrieved data.
}
\description{
Retrieves data from the TRON network page by page using the Tronscan API
}
\examples{
tx_data <- tronr:::run_paginated_tronscan_query(
  base_url = "https://apilist.tronscan.org/",
  path = c("api", "transaction"),
  params = list(
    sort = "-timestamp",
    count = "true",
    limit = 25,
    block = 26808690
  )
)
print(tx_data)
}
