% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_unix_timestamp.R
\name{from_unix_timestamp}
\alias{from_unix_timestamp}
\title{Convert to Unix timestamp}
\usage{
from_unix_timestamp(ts, tz = "UTC")
}
\arguments{
\item{ts}{A Unix timestamp, \emph{including milliseconds}. Either a numeric or
a character value.}

\item{tz}{Character value corresponding to the timezone of the returned
datetime value. Defaults to \code{UTC}.}
}
\value{
A \code{POSIXct} datetime value in the format \verb{\%Y-\%m-\%d \%H:\%M:\%S}.
}
\description{
Converts a Unix timestamp to a POSIXct datatime value
}
\examples{
ts <- 60 * 60 * 24 * 1000
from_unix_timestamp(ts, tz = "UTC")
}
