% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_balance.R
\name{get_account_balance}
\alias{get_account_balance}
\title{Get account balance}
\usage{
get_account_balance(address, max_attempts = 3L)
}
\arguments{
\item{address}{(character): address of the account of interest,
either in \code{base58check} or \code{hex} format.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble with the following columns:
\itemize{
\item \code{request_time} (POSIXct, UTC timezone): date and time of the request;
\item \code{address} (character): account address (in \code{base58check} format);
\item \code{name} (character): name of the account (\code{NA} if absent);
\item \code{total_transaction_count}: total number of transactions recorded for
the account as of \code{request_time};
\item \code{bandwidth} (list): contains a tibble with several variables describing
the energy and bandwidth usage of the account;
\item \code{trx_balance} (double): currently available (i.e. not-frozeb) balance of
Tronix (TRX);
\item \code{n_trc20} (integer): number of unique TRC-20 tokens currently held by the
account (\code{0} if absent);
\item \code{trc20} (list or \code{NA} if absent): contains a tibble with \code{n_trc20} rows
and several variables describing the TRC-20 tokens held by the account;
\item \code{n_trc10} (integer): number of unique TRC-10 tokens currently held by the
account (\code{0} if absent);
\item \code{trc10} (list or \code{NA} if absent): contains a tibble with \code{n_trc10} rows
and several columns describing the TRC-10 assets held by the account.
}
}
\description{
Retrieves the current token and bandwidth balances of an account
}
\examples{
r <- get_account_balance("TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux")
print(r)
}
\seealso{
\code{\link[=get_account_trc10_balance]{get_account_trc10_balance()}}.
}
