% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_trc10_balance.R
\name{get_account_trc10_balance}
\alias{get_account_trc10_balance}
\title{Get TRC-10 token balances}
\usage{
get_account_trc10_balance(
  address,
  only_confirmed = NULL,
  detailed_info = FALSE,
  max_attempts = 3L
)
}
\arguments{
\item{address}{(character): address of the account of interest,
either in \code{base58check} or \code{hex} format.}

\item{only_confirmed}{(boolean or \code{NULL}): if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for confirmed
transactions are returned.}

\item{detailed_info}{(boolean): if \code{FALSE} (default),
only basic information about the TRC-10 token assets is returned.
If \code{TRUE}, an extended information is returned.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble with the following columns:
\itemize{
\item \code{request_time} (POSIXct, UTC timezone): date and time when the API
request was made;
\item \code{address} (character): account address, in \code{base58} format;
\item \code{n_trc10} (double): number of unique TRC-10 assets held by \code{address};
\item \code{trc10_balance} (list): contains a tibble with \code{n_trc20} rows and
several attributes of the TRC-10 assets held by \code{address}. The actual
content of this tibble will depend on the argument \code{detailed_info}
(see \code{\link[=get_asset_by_id]{get_asset_by_id()}} for details).
}
}
\description{
Returns information on the current TRC-10 assets held by an account
}
\details{
TRC-10 is a technical standard used by system contracts to
implement tokens. If an account holds no TRC-10 tokens (\code{n_trc10 = 0}),
the \code{trc10_balance} column in the resultant tibble will contain an
\code{NA} value.
}
\examples{
r <- get_account_trc10_balance("TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux")
print(r)
}
