% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_trx_balance.R
\name{get_account_trx_balance}
\alias{get_account_trx_balance}
\title{Get TRX balance}
\usage{
get_account_trx_balance(address, only_confirmed = FALSE, max_attempts = 3L)
}
\arguments{
\item{address}{A character value - address of the account of interest, in
\code{base58} (starts with \code{T}) or  \code{hex} (starts with \code{41}) format.}

\item{only_confirmed}{A boolean value. If \code{TRUE}, account balance will be
returned as of the latest confirmed block, otherwise as of the
latest unconfirmed one. Defaults to \code{FALSE}.}

\item{max_attempts}{A non-zero, positive integer specifying the maximum
number of attempts to call the API if the first attempt fails
(i.e. its call status is different from \code{200}). Additional attempts are
implemented with an exponential backoff. Defaults to 3.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{request_time}: date and time  (UTC timezone) when the API
request was made;
\item \code{address}: a character value indicating the account address
(in \code{hex} format);
\item \code{trx_balance}: a character value.
}
}
\description{
Returns the current TRX balance of an account
}
\details{
All balances are presented with a precision of 6. This means
that a balance returned by this function needs to be divided by 6
(after converting to \code{as.numeric}) to obtain the actual value.
}
\examples{
r <- get_account_trx_balance("TQjaZ9FD473QBTdUzMLmSyoGB6Yz1CGpux")
print(r)

}
