% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_by_id.R
\name{get_asset_by_id}
\alias{get_asset_by_id}
\title{Get description of a TRC-10 asset}
\usage{
get_asset_by_id(
  asset_id,
  only_confirmed = NULL,
  detailed_info = FALSE,
  max_attempts = 3L
)
}
\arguments{
\item{asset_id}{(character): ID of the TRC-10 token of interest,
presented either as a set of numbers (e.g., \code{1002762}) or as
address of the owner who issued that token (in \code{base58hex} or
\code{hex} format). Using either of this IDs is possible because there
is a 1:1 relationship between them, i.e. an account is only
allowed to issue a single TRC-10 token.}

\item{only_confirmed}{(boolean or \code{NULL}): if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for confirmed
transactions are returned.}

\item{detailed_info}{(boolean): if \code{FALSE} (default),
only basic information about the TRC-10 token assets is returned.
If \code{TRUE}, an extended information is returned.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble, whose content depends on the \code{detailed_info} argument. If
\code{detailed_info = FALSE} (default), the tibble will have the following 5
columns:
\itemize{
\item \code{asset_id} (character): same as argument \code{asset_id};
\item \code{owner_address} (character): address of the asset issuer, in \code{base58}
format;
\item \code{abbr} (character): abbreviated name of the asset;
\item \code{asset_name} (character): full name of the asset
\item \code{precision} (integer): precision used to present the asset's balance
(e.g., if it is 6, then one needs to divide the returned balance by 1
million to obtain the actual balance of that asset).
}

If \code{detailed_info = TRUE}, the returned tibble will have the same 5 columns
as above and the following additional columns:
\itemize{
\item \code{description} (character): a free-text field describing the asset;
\item \code{url} (character): URL of the token's project;
\item \code{total_supply} (character): total issued amount of tokens;
\item \code{num} (character): amount of the asset's tokens that one can buy
with \code{trx_num} TRX tokens (see next point);
\item \code{trx_num} (character): amount of TRX tokens that is required to buy \code{num}
tokens of the asset (thus, \code{num / num_trx} is the asset's price during
its ICO);
\item \code{ico_start_time} (POSIXct, UTC timezone): date and time of the asset's ICO
start;
\item \code{ico_end_time} (POSIXct, UTC timezone): date and time of the asset's ICO
end.
}
}
\description{
Returns information about a TRC-10 token based on its ID
}
\details{
TRC-10 are tokens issued by the system contract (as opposed to
TRC-20, which are issued by smart contracts). See
\href{https://developers.tron.network/docs/trc10}{official documentation}
for details.
}
\examples{
r <- get_asset_by_id(asset_id = "1002762", detailed_info = TRUE)
print(r)
}
