% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trc20_tx_info_by_account_address.R
\name{get_trc20_tx_info_by_account_address}
\alias{get_trc20_tx_info_by_account_address}
\title{Get TRC-20 transactions for an account}
\usage{
get_trc20_tx_info_by_account_address(
  address,
  only_confirmed = NULL,
  only_unconfirmed = NULL,
  only_to = FALSE,
  only_from = FALSE,
  min_timestamp = 0,
  max_timestamp = NULL,
  contract_address = NULL,
  max_attempts = 3L
)
}
\arguments{
\item{address}{(character): address of the account of interest,
either in \code{base58check} or \code{hex} format.}

\item{only_confirmed}{(boolean or \code{NULL}): if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for confirmed
transactions are returned.}

\item{only_unconfirmed}{(boolean or \code{NULL}): if \code{NULL}
(default) or \code{FALSE}, results are returned for both confirmed and
unconfirmed transactions. If \code{TRUE}, only results for
unconfirmed transactions are returned. Cannot be used
simultanously with the \code{only_confirmed} argument.}

\item{only_to}{(boolean, defaults to \code{FALSE}): if \code{TRUE}, only
inbound transactions are returned.}

\item{only_from}{(boolean, defaults to \code{FALSE}): if \code{TRUE},
only outbound transactions are returned.}

\item{min_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
beginning of the period of interest (inclusive). Defaults to 0.}

\item{max_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
end of the period of interest (inclusive).}

\item{contract_address}{(character): contract address of
a specific TRC-20 token of interest, either in \code{base58check} or
\code{hex} format.}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call the URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A tibble where each row corresponds to one transaction.
This tibble contains the following columns:
\itemize{
\item \code{tx_id} (character) - transation ID;
\item \code{tx_type} (character) - transation type (e.g., \code{"Transfer"});
\item \code{block_timestamp} (POSIXct);
\item \code{from_address} (character, \code{base58check}-formatted);
\item \code{to_address} (character, \code{base58check}-formatted);
\item \code{trc20_symbol} (character) - abbreviated name of the TRC-20 token;
\item \code{trc20_name} (character) - common name of the TRC-20 token;
\item \code{trc20_contract_address} (character, \code{base58check}-formatted);
\item \code{precision} (double) - precision of the \code{amount} values;
\item \code{amount} (double) - transfered amount.
}

If no TRC-20 transactions are found for the specified combination of query
parameters, nothing (\code{NULL}) is returned, with a console message
\code{"No data found"}.
}
\description{
Returns a list of TRC-20 transactions associated with an address
}
\details{
Some addresses have a very high load of transactions going
through them. Users are, therefore, advised to choose \code{min_timestamp} and
\code{max_timestamp} wisely as the TronGrid API may deny calls that
attempt to retrieve a large amount of data in one go. Such requests will
fail with status \code{404} and an error message
\code{"Exceeds the maximum limit, please change your query time range"}).
Chunking the time range of interest into smaller periods may help,
however users will have to implement their own logic for this.
}
\examples{
tx_df <- get_trc20_tx_info_by_account_address(
  address = "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX",
  min_timestamp = "1604188800000",
  max_timestamp = "1604189100000"
)
print(tx_df)
}
