% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_info_by_account_address.R
\name{get_tx_info_by_account_address}
\alias{get_tx_info_by_account_address}
\title{Get transactions for a given account}
\usage{
get_tx_info_by_account_address(
  address,
  only_confirmed = NULL,
  only_unconfirmed = NULL,
  only_to = FALSE,
  only_from = FALSE,
  min_timestamp = 0,
  max_timestamp = NULL,
  limit = 200L,
  max_attempts = 3L
)
}
\arguments{
\item{address}{A character value - address of the account of interest, in
\code{base58} (starts with \code{T}) or \code{hex} (starts with \code{41}) format.}

\item{only_confirmed}{A boolean value. If \code{NULL} (default) or \code{FALSE},
both confirmed and unconfirmed transactions are returned. If \code{TRUE},
only confirmed transactions are returned. Cannot be used simultanously
with the \code{only_unconfirmed} argument (see next).}

\item{only_unconfirmed}{A boolean value. If \code{NULL} (default) or \code{FALSE},
both confirmed and unconfirmed transactions are returned. If \code{TRUE},
only unconfirmed transactions are returned. Cannot be used simultanously
with the \code{only_confirmed} argument.}

\item{only_to}{A boolean value (defautls to \code{FALSE}). If \code{TRUE}, only
inbound transactions are returned.}

\item{only_from}{A boolean value (defautls to \code{FALSE}). If \code{TRUE}, only
inbound transactions are returned.}

\item{min_timestamp}{A Unix timestamp (\emph{including milliseconds}), presented
as either a numeric or a character value. Defines the beginning of the
period to retrieve the transactions from.}

\item{max_timestamp}{A Unix timestamp (\emph{including milliseconds}), presented
as either a numeric or a character value. Defines the end of the
period to retrieve the transactions from.}

\item{limit}{An integer - number of transactions per page. Defaults to 200.
Maximum accepted value is 200.}

\item{max_attempts}{A non-zero, positive integer specifying the maximum
number of additional attempts to call the API if the first attempt fails
(i.e. its call status is different from \code{200}). Additional attempts are
implemented with an exponential backoff. Defaults to 3.}
}
\value{
A nested tibble where each row corresponds to one transaction.
This tibble contains the following columns:
\itemize{
\item \code{tx_id} - transation ID;
\item \code{tx_type} - transation type (see \href{https://tronscan-org.medium.com/tronscan-class-transaction-b6b3ea681e43}{here}
and \href{https://tronscan-org.medium.com/tronscan-class-transaction-b6b3ea681e43}{here}
for a list of all possible values and further details);
\item \code{tx_result} - transation result;
\item \code{net_usage};
\item \code{net_fee};
\item \code{energy_usage};
\item \code{energy_fee};
\item \code{block_number};
\item \code{block_timestamp};
\item \code{raw_data} - a list column where each element contains a tibble with
additional transaction attributes (the actual structure of this
tibble will depend on \code{tx_type}, but among other things it will typically
contain \code{from_address}, \code{to_address} and transaction \code{timestamp});
\item \code{internal_tx} - a list column where each element contains a list with
attributes of the internal transactions triggered as part of \code{tx_id} (the
actual structure of this list will depend on \code{tx_type}), or \code{NA} if no
internal transactions were triggered.
}
}
\description{
Returns various bits of information about the transactions of a given account
}
\details{
Some accounts may have a very high load of transactions going
through them. Users are, therefore, advised to choose \code{min_timestamp} and
\code{max_timestamp} wisely as the TronGrid API will not return more than
10000 transactions in one request. If the number of transactions
exceeds this limit, the the request will fail with status \code{404} and an
error message
\code{"Exceeds the maximum limit, please change your query time range"}). If
data are to be retrieved for a large time range, users are advised
to implement their own logic for querying the network (e.g.,
splitting that range into smaller chunks and then combining the results).
}
\examples{
tx_df <- get_tx_info_by_account_address(
                         address = "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX",
                         only_confirmed = TRUE,
                         only_from = TRUE,
                         min_timestamp = "1577836800000",
                         max_timestamp = "1577838600000",
                         limit = 10L
                         )
print(tx_df)

}
