% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_info_by_account_address.R
\name{get_tx_info_by_account_address}
\alias{get_tx_info_by_account_address}
\title{Get transactions for a given account}
\usage{
get_tx_info_by_account_address(
  address,
  only_confirmed = NULL,
  only_unconfirmed = NULL,
  only_to = FALSE,
  only_from = FALSE,
  min_timestamp = 0,
  max_timestamp = NULL,
  limit = 10L,
  search_internal = TRUE,
  max_attempts = 3L
)
}
\arguments{
\item{address}{A character value - address of the account of interest, in
\code{base58} (starts with \code{T}) or \code{hex} (starts with \code{41}) format.}

\item{only_confirmed}{A boolean value. If \code{NULL} (default) or \code{FALSE},
both confirmed and unconfirmed transactions are returned. If \code{TRUE},
only confirmed transactions are returned. Cannot be used simultanously
with the \code{only_unconfirmed} argument (see next).}

\item{only_unconfirmed}{A boolean value. If \code{NULL} (default) or \code{FALSE},
both confirmed and unconfirmed transactions are returned. If \code{TRUE},
only unconfirmed transactions are returned. Cannot be used simultanously
with the \code{only_confirmed} argument.}

\item{only_to}{A boolean value (defautls to \code{FALSE}). If \code{TRUE}, only
inbound transactions are returned.}

\item{only_from}{A boolean value (defautls to \code{FALSE}). If \code{TRUE}, only
inbound transactions are returned.}

\item{min_timestamp}{A Unix timestamp (\emph{including milliseconds}), presented
as either a numeric or a character value. Defines the beginning of the
period to retrieve the transactions from.}

\item{max_timestamp}{A Unix timestamp (\emph{including milliseconds}), presented
as either a numeric or a character value. Defines the end of the
period to retrieve the transactions from.}

\item{limit}{An integer - number of transactions per page. Defaults to 20.
Maximum accepted value is 200.}

\item{search_internal}{A boolean value. If \code{TRUE} (default), query params are
applied to both normal and internal transactions. If \code{FALSE}, query
params are only applied to normal transactions.}

\item{max_attempts}{A non-zero, positive integer specifying the maximum
number of additional attempts to call the API if the first attempt fails
(i.e. its call status is different from \code{200}). Additional attempts are
implemented with an exponential backoff. Defaults to 3.}
}
\value{
A data frame where each row corresponds to one transaction.
}
\description{
Returns various bits of information about the transactions of a given account
}
\details{
As some accounts may have a very high load of transactions going
through them, this function will only return data for a maximum of 1 day.
This behaviour is enforced even if the difference between the requested
\code{max_timestamp} and \code{min_timestamp} is larger than 1 day. The user is
advised to implement their own logic if they need to retrieve data
covering a larger period of time.
}
\examples{
tx_df <- get_tx_info_by_account_address(
                         address = "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX",
                         only_confirmed = TRUE,
                         only_from = TRUE,
                         min_timestamp = "1577836800000",
                         max_timestamp = "1577838600000",
                         limit = 10L
                         )
print(tx_df)

}
