% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx_info_by_account_address.R
\name{get_tx_info_by_account_address}
\alias{get_tx_info_by_account_address}
\title{Get transactions for an account}
\usage{
get_tx_info_by_account_address(
  address,
  min_timestamp = 0,
  max_timestamp = NULL,
  add_contract_data = TRUE,
  max_attempts = 3L
)
}
\arguments{
\item{address}{(character): address of the account of interest,
either in \code{base58check} or \code{hex} format.}

\item{min_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
beginning of the period of interest (inclusive). Defaults to 0.}

\item{max_timestamp}{(numeric or character): a Unix
timestamp (\emph{including milliseconds}), which defines the
end of the period of interest (inclusive).}

\item{add_contract_data}{(boolean): if \code{TRUE} (default),
adds column \code{contract_data} to the resultant tibble (see Value).}

\item{max_attempts}{(integer, positive): specifies the
maximum number of additional attempts to call a URL if the
first attempt fails (i.e. its call status is different from
\code{200}). Additional attempts are implemented with an exponential
backoff. Defaults to \code{3}.}
}
\value{
A nested tibble where each row corresponds to one transaction.
A detailed description of the content of this tibble can be found in the
help file for \code{\link[=get_tx_info_by_id]{get_tx_info_by_id()}}.
}
\description{
Returns a list of transactions associated with an account and their attributes
}
\details{
As the number of transactions associated with some TRON addresses
can be very high, users are advised to choose \code{min_timestamp} and
\code{max_timestamp} wisely. If the query range is too large, the maximum
number of transactions returned by the underlying Tronscan API will be
\emph{capped} at 2000. Chunking the time range of interest into smaller
periods can help to avoid gaps in data in such cases. However, users
would have to implement their own logic for that.
}
\examples{
tx_df <- get_tx_info_by_account_address(
  address = "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX",
  min_timestamp = "1577836800000",
  max_timestamp = "1577838600000"
)
print(tx_df)

}
