% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_application_json.R
\name{is_application_json}
\alias{is_application_json}
\title{Validate JSON objects}
\usage{
is_application_json(response)
}
\arguments{
\item{response}{A response object as returned by the \code{httr}'s request verbs
(e.g., \code{httr::GET} or \code{httr::POST}).}
}
\value{
A boolean value (\code{TRUE} if the API call returned JSON data).
}
\description{
Checks if an API call returned JSON-formatted data.
}
\keyword{internal}
