% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_events_info.R
\name{parse_events_info}
\alias{parse_events_info}
\title{Parse event attributes}
\usage{
parse_events_info(info)
}
\arguments{
\item{info}{A list returned as a result of calling the
\href{https://developers.tron.network/reference#events}{events-related methods}
of the TronGird API.}
}
\value{
A nested tibble with one row and the following columns:
\itemize{
\item \code{tx_id} (character) - ID of the transaction that a given event is part of;
\item \code{block_number} (character);
\item \code{block_timestamp} (POSIXct, UTC timezone);
\item \code{contract_address} (character) adress of the contract that performed the
transaction of interest;
\code{event_name} (character) - possible values of this column will depend on
the nature of the transaction of interest;
\code{event_data} (list) - each element of this list contains a tibble with
additional attributes of the event. The exact content of this tibble will
be contract- and event-specific. Thus, very little processing is done with
these data, except for removing redundant attributes and converting all
addresses to \code{base58} format.
}
}
\description{
Converts a list with event attributes into a nested tibble
}
\examples{
address <- "TKttnV3FSY1iEoAwB4N52WK2DxdV94KpSd"
min_timestamp <- "1576317786000"
max_timestamp <- "1576317996000"
query_params <- list(min_block_timestamp = min_timestamp,
                     max_block_timestamp = max_timestamp)
url <- tronr::build_get_request(base_url = "https://api.trongrid.io",
                                path = c("v1", "contracts",
                                address, "events"),
                                query_parameters = query_params)
r <- tronr::api_request(url = url, max_attempts = 3L)
parse_events_info(r$data[[1]])

}
