% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tx_info.R
\name{parse_tx_info}
\alias{parse_tx_info}
\title{Parse transaction attributes}
\usage{
parse_tx_info(info)
}
\arguments{
\item{info}{A list returned as a result of calling the
"get transaction info by account address" method.}
}
\value{
A nested tibble where each row corresponds to one transaction.
}
\description{
Converts a list with transaction attributes into a nested tibble
}
\details{
This is a non-public function that is used to simplify syntax of
the \code{get_tx_info_by_account_address()} function.
}
\examples{
query_params <- list(only_confirmed = tolower(TRUE),
                               only_to = tolower(TRUE),
                               min_timestamp = "1577836800000",
                               max_timestamp = "1577838600000",
                               limit = 100L)
address <- "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX"
url <- tronr::build_get_request(base_url = "https://api.trongrid.io",
                                path = c("v1", "accounts",
                                address, "transactions"),
                                query_parameters = query_params)
r <- tronr::api_request(url = url, max_attempts = max_attempts)
tx_info <- r$data
tx_info_parsed <- parse_tx_info(tx_info)
print(tx_info_parsed)

}
