% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tx_info.R
\name{parse_tx_info}
\alias{parse_tx_info}
\title{Parse transaction attributes}
\usage{
parse_tx_info(info)
}
\arguments{
\item{info}{(list): a non-empty, named list returned as a result of calling the
\href{https://github.com/tronscan/tronscan-frontend/blob/dev2019/document/api.md}{\verb{/api/transaction-info}}
endpoint of the Tronscan API.}
}
\value{
A nested tibble with one row and several columns that contain
transaction attributes. See \code{\link[=get_tx_info_by_id]{get_tx_info_by_id()}} for details.
}
\description{
Converts a list with transaction attributes into a nested tibble
}
\examples{
id <- "dca447279bc2fd3c10325e442746f9a42938e25bac33bc277b3e7720027aaaf2"
url <- build_get_request(
  base_url = "https://apilist.tronscan.org/",
  path = c("api", "transaction-info"),
  query_parameters = list(hash = id)
)
r <- api_request(url = url, max_attempts = 3L)
result <- parse_tx_info(info = r)
print(result)
}
