% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_unix_timestamp.R
\name{to_unix_timestamp}
\alias{to_unix_timestamp}
\title{Convert to Unix timestamp}
\usage{
to_unix_timestamp(datetime, tz = "UTC")
}
\arguments{
\item{datetime}{Either a \code{POSIXct} datetime value or a character value that
can be coerced into such a \code{POSIXct} datetime value. Expected format:
\verb{\%Y-\%m-\%d \%H:\%M:\%S}. If \code{datetime} is a character value, then hours,
minutes, and seconds must be provided explicitly (so, for example,
a value of \code{"2010-01-01 00:00:00"} will work, while \code{"2010-01-01"} will
throw an error).}

\item{tz}{Character value corresponding to the timezone of \code{datetime}.
Defaults to \code{UTC}.}
}
\value{
A character value that corresponds to a Unix timestamp.
}
\description{
Converts a POSIX datetime to Unix timestamp
}
\details{
The TronGrid API commands expect Unix timestamps with a
millisecond-level precision. This function will therefore automatically
add 3 more zeros after converting \code{datetime} into a Unix timestamp.
}
\examples{
dt <- "2019-01-01 00:00:10"
to_unix_timestamp(datetime = dt, tz = "UTC")
}
